Profile editor - Added tools and layer visualizers

+++ src/Avalonia/Artemis.UI/Screens/ProfileEditor/Panels/VisualEditor/Visualizers/IVisualizerViewModel.cs
@@ -2,7 +2,7 @@
 
 public interface IVisualizerViewModel
 {
-    int X { get; }
-    int Y { get; }
+    double X { get; }
+    double Y { get; }
     int Order { get; }
 }
